requirements.txt
fastapi==0.95.1 # FAST API framework
uvicorn==0.22.0 # ASGI server
pydantic==1.10.7 # Data validation and settings management using python type annotations
sqlalchemy==2.0.12 # ORM for database
python-dotenv==1.0.0 # .env file support
